TEST: move classifier fixtures to conftest.py
